fix: address code review issues - regex backtracking, capability scope, thinking detection
- Change capBlockRe from [^]*? to [\s\S]*? to avoid regex backtracking
- Scope capability detection to dedicated capabilities section to avoid false positives in descriptions
- Add thinking-specific check to isThinkingInternalServerError to avoid false positives
- Create shared formatting.ts with formatXmlLikeResponseForDisplay to eliminate duplication
